requirements.txt
tensorflow
opencv-python
numpy
python-telegram-bot
requests
pillow
scikit-learn
